// src/Core/taxi/dto/taxi-pay.dto.ts
export class TaxiPayDto {
  taxiInquiryId: string;
  amount: number;
  mobile: string;
  entityId: string;
  amountAfterDiscount: number;
  terminalID: number;
  instituteId: number;
  customerId: string;
  nationalCode: string;
  Description: string;
  transactionDate?: string;
  receiveDate?: string;
  pan: string;
  devicetype: string;
  user?: string;
}